WIP: USE PRISMA - (build will fail)
This is a huuuge PR, I'm committing most of it just for safety, this has been living on my machine for a while and I don't want to loose it one day.
Working on not self-hosting Postgres and using Prisma.
Removing sagas from Store, rewriting some of the reducers, making user everything that needs authentication is authenticated.
Slight API restructure and more abstraction so Prisma can be replaced more easily in the future if needed.